% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{char_in_string}
\alias{char_in_string}
\title{char_in_string}
\usage{
char_in_string(c, str)
}
\arguments{
\item{c}{character}

\item{str}{string}
}
\value{
bool: true if char is in string
}
\description{
checks if a character is contained in a string
}
