% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clean_syntax}
\alias{clean_syntax}
\title{clean_syntax}
\usage{
clean_syntax(syntax)
}
\arguments{
\item{syntax}{lavaan style syntax}
}
\value{
vector of strings with cleaned syntax
}
\description{
takes in a lavaan style syntax and removes comments, white space, etc.
}
