% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_algebra_elements.R
\name{extract_algebra_elements}
\alias{extract_algebra_elements}
\title{extract_algebra_elements}
\usage{
extract_algebra_elements(mxAlgebra_formula, extracted = c())
}
\arguments{
\item{mxAlgebra_formula}{formula embedded in mxAlgebra}

\item{extracted}{used in recursive function calls; don't set this manually}
}
\value{
vector with names of variables and parameters used in the function call
}
\description{
extract all variables/parameters from an mxAlgebra
}
\keyword{internal}
