% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_model_name}
\alias{find_model_name}
\title{find_model_name}
\usage{
find_model_name(syntax)
}
\arguments{
\item{syntax}{lavaan like syntax}
}
\value{
vector with (1) model name and (2) model syntax
}
\description{
checks for a model name in the syntax
}
\keyword{internal}
