% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parameter_table_rcpp}
\alias{parameter_table_rcpp}
\title{parameter_table_rcpp}
\usage{
parameter_table_rcpp(
  syntax,
  add_intercept,
  add_variance,
  add_exogenous_latent_covariances,
  add_exogenous_manifest_covariances,
  scale_latent_variance,
  scale_loading
)
}
\arguments{
\item{syntax}{lavaan like syntax}

\item{add_intercept}{should intercepts for manifest variables be automatically added?}

\item{add_variance}{should variances for all variables be automatically added?}

\item{add_exogenous_latent_covariances}{should covariances between exogenous latent variables be
added automatically?}

\item{add_exogenous_manifest_covariances}{should covariances between exogenous manifest variables be
added automatically?}

\item{scale_latent_variance}{should variances of latent variables be set to 1?}

\item{scale_loading}{should the first loading of each latent variable be set to 1?}
}
\value{
parameter table
}
\description{
creates a parameter table from a lavaan like syntax
}
