% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parameter_table_to_Rcpp}
\alias{parameter_table_to_Rcpp}
\title{parameter_table_to_Rcpp}
\usage{
parameter_table_to_Rcpp(equations)
}
\arguments{
\item{equations}{vector with equations}
}
\value{
parameter table
}
\description{
creates a parameter table from equations
}
