% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{parameters}
\usage{
parameters(mxMod)
}
\arguments{
\item{mxMod}{mxModel object}
}
\value{
vector with parameter estimates
}
\description{
Returns the parameter estimates of an mxModel. Wrapper
for omxGetParameters
}
