% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxsem_group_by.R
\name{print.multi_group_parameters}
\alias{print.multi_group_parameters}
\title{print the multi_group_parameters}
\usage{
\method{print}{multi_group_parameters}(x, ...)
}
\arguments{
\item{x}{object from summarize_multi_group_model}

\item{...}{not used}
}
\description{
print the multi_group_parameters
}
