% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{split_string_all}
\alias{split_string_all}
\title{split_string_all}
\usage{
split_string_all(str, at)
}
\arguments{
\item{str}{string to be splitted}

\item{at}{where to split the string at}
}
\value{
vector of strings
}
\description{
splits a string
}
\keyword{internal}
