% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicode_arrows.R
\name{unicode_directed}
\alias{unicode_directed}
\title{unicode_directed}
\usage{
unicode_directed()
}
\value{
returns unicode for directed arrows
}
\description{
this function returns the unicode for directed arrows
}
