% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicode_arrows.R
\name{unicode_undirected}
\alias{unicode_undirected}
\title{unicode_undirected}
\usage{
unicode_undirected()
}
\value{
returns unicode for undirected arrows
}
\description{
this function returns the unicode for undirected arrows
}
