% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parameter_table_rcpp}
\alias{parameter_table_rcpp}
\title{parameter_table_rcpp}
\usage{
parameter_table_rcpp(
  syntax,
  add_intercept,
  add_variance,
  scale_latent_variance,
  scale_loading
)
}
\arguments{
\item{syntax}{lavaan like syntax}

\item{add_variance}{should variances for all variables be automatically added?}

\item{add_intercepts}{should intercepts for manifest variables be automatically added?}
}
\value{
parameter table
}
\description{
creates a parameter table from a lavaan like syntax
}
